Refactor Vaults page to split components and be easier to understand
