+++ apps/vaults/components/detail/VaultStrategiesSection.tsx
@@ -6,9 +6,9 @@ import type { TSortDirection } from '@lib/types'
 import { cl, formatCounterValue, formatPercent, toNormalizedBN } from '@lib/utils'
 import type { TYDaemonVault, TYDaemonVaultStrategy } from '@lib/utils/schemas/yDaemonVaultsSchemas'
 import { ALL_VAULTSV3_KINDS_KEYS } from '@vaults/constants'
-import type { TPossibleSortBy } from '@vaults/shared/hooks/useSortVaults'
-import { useSortVaults } from '@vaults/shared/hooks/useSortVaults'
-import { useQueryArguments } from '@vaults/shared/hooks/useVaultsQueryArgs'
+import type { TPossibleSortBy } from '@vaults/hooks/useSortVaults'
+import { useSortVaults } from '@vaults/hooks/useSortVaults'
+import { useQueryArguments } from '@vaults/hooks/useVaultsQueryArgs'
 import type { ReactElement } from 'react'
 import { lazy, Suspense, useMemo } from 'react'
 import { VaultsListHead } from './VaultsListHead'
