+++ apps/vaults/components/list/VaultHoldingsAmount.tsx
