+++ apps/vaults/components/list/VaultsListRow.tsx
@@ -16,7 +16,7 @@ import { getNetwork } from '@lib/utils/wagmi'
 import type { TVaultChartTimeframe } from '@vaults/components/detail/VaultChartsSection'
 import { type TVaultForwardAPYVariant, VaultForwardAPY } from '@vaults/components/table/VaultForwardAPY'
 import { VaultHoldingsAmount } from '@vaults/components/table/VaultHoldingsAmount'
-import { deriveListKind } from '@vaults/shared/utils/vaultListFacets'
+import { deriveListKind } from '@vaults/utils/vaultListFacets'
 import type { ReactElement } from 'react'
 import { lazy, Suspense, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router'
