--- apps/vaults/shared/index.ts
@@ -1,3 +1,0 @@
-export { VaultListEmptyExternalMigration, VaultsListEmpty } from './components/list/VaultsListEmpty'
-export { type TPossibleSortBy, useSortVaults } from './hooks/useSortVaults'
-export { type TQueryArgs, useQueryArguments } from './hooks/useVaultsQueryArgs'
