+++ apps/vaults/utils/vaultTypeCopy.ts
