+++ apps/vaults/utils/vaultTypeUtils.ts
